chore: fix tests & cleanup
